database/requirements.txt
fastapi==0.109.2
httpx==0.26.0
pydantic==2.6.1
python-dotenv==1.0.1
uvicorn==0.27.0
PyJWT==2.3.0
SQLAlchemy==2.0.27
bcrypt==4.1.2
passlib==1.7.4